# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Levels Derm Clinic API",
    "version": "1.0.0",
    "description": "API for cosmetic clinic management system with Arabic support"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "patient_id": {
            "type": "string",
            "example": "PAT000001"
          },
          "national_id": {
            "type": "string",
            "example": "1234567890"
          },
          "first_name": {
            "type": "string",
            "example": "أحمد"
          },
          "last_name": {
            "type": "string",
            "example": "محمد علي"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ahmed@example.com"
          },
          "phone": {
            "type": "string",
            "example": "0501234567"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "example": "1990-01-01"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female"],
            "example": "male"
          },
          "address": {
            "type": "string",
            "example": "الرياض، حي النرجس"
          },
          "city": {
            "type": "string",
            "example": "الرياض"
          },
          "state": {
            "type": "string",
            "example": "منطقة الرياض"
          },
          "country": {
            "type": "string",
            "example": "السعودية"
          },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "suspended"],
            "example": "active"
          },
          "visit_count": {
            "type": "integer",
            "example": 5
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/patients": {
      "get": {
        "tags": ["Patients"],
        "summary": "Get all patients",
        "description": "Get list of all patients with pagination and search",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 15
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for patient name or phone",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم جلب قائمة المرضى بنجاح"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Patient"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}": {
      "get": {
        "tags": ["Patients"],
        "summary": "Get patient details",
        "description": "Get detailed information about a specific patient",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Patient ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "تم جلب بيانات المريض بنجاح"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Patient"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}